refactoring: CardSelector -> CompanySelection 이름 변경
